requirements.txt
paramiko>=2.9.0
PyYAML>=6.0
click>=8.0.0
rich>=12.0.0
colorlog>=6.6.0
tabulate>=0.9.0
pydantic>=1.8.0 